+++ .gitignore
@@ -1,2 +1,4 @@
 .idea
 **/node_modules
+.vscode
+.expo
