+++ bot/src/app/utils/Helpers.js
@@ -4,23 +4,6 @@ export function isObjectEmpty(obj) {
   return Object.keys(obj).length === 0 && obj.constructor === Object;
 }
 
-export function getQueryStringValue(key) {
-  const urlParams = new URLSearchParams(window.location.search);
-  return urlParams.get(key);
-}
-
-export function pageHasNodeOfClass(className) {
-  return document.getElementsByClassName(className).length > 0;
-}
-
-export function hasStringInPath(value) {
-  return location.pathname.substring(1).split('/').filter(x => !!x && x === value).length > 0;
-}
-
-export function pathCount() {
-  return location.pathname.substring(1).split('/').length;
-}
-
 export function openAtcTab(category, keywords, color) {
   let url = `http://supremenewyork.com/shop/all/${category}?atc-kw=${keywords.join(';')}`;
   if (color) {
