+++ bot/src/extension/processors/baseProcessor.js
@@ -1,0 +1,7 @@
+export default class BaseProcessor {
+  constructor(preferences, sizings, billing) {
+    this.preferences = preferences;
+    this.sizings = sizings;
+    this.billing = billing;
+  }
+}
